refactor(routes): extract agents-channels.ts (managed settings + channels setup/access)
Moves channel-related handlers out of agents.ts into a new agents-channels.ts:
- managedSettingsPath, isManagedSettingsReady, getManagedSettingsSudoCommand,
  setAgentEnabledPlugins, resetAgentEnabledPlugins (managed settings block)
- Slack manifest + smoke-test, channel reconnect + health
- Channel setup wizard (POST/DELETE) for all providers
- Pending/approve, allowed list/remove, invites create/revoke
- Channel requests list/approve/deny
- Auth init (/login scraper)

agents.ts delegates with tryHandleAgentsChannels(ctx) and re-exports the
managed-settings functions for backward compatibility with existing tests.
Drops 11 dead imports from agents.ts post-extraction.
agents.ts: 1435 -> 741 lines; agents-channels.ts: 713 lines new.
